pyproject.toml
[project]
name = "gemini-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "google-genai>=1.26.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.28.0",
]

[tool.uv]
dev-dependencies = []
